refactor: refactor repository for filter

+++ domain/model/src/main/java/com/crediya/solicitudes/model/PageResponse.java
@@ -3,10 +3,10 @@ package com.crediya.solicitudes.model;
 import java.util.List;
 
 public record PageResponse<T>(
-        int page,
-        int size,
-        int totalElements,
-        int totalPages,
+        Integer page,
+        Integer size,
+        Long totalElements,
+        Integer totalPages,
         List<T> content
 ) {
 }
